updated team page
